webadmin: Add nic via guideme sets the wrong nic name (#846997)
https://bugzilla.redhat.com/846997

Sometimes, when adding nic via guide me, the nic is created with the
default parameters and not with the parameters the user specified in the
add nic dialog.
The reason - In guide me model- when adding a nic there are async queries
for nics and networks. When the nics and networks arrives from the backend
(nic!=null && networks!=null) the adding procedure should continue and
open the add-nic dialog.
In case the networks query is ready before the nics query- the adding
procedure continue (because nics were already initialized before) and
opens the dialog. Then when the nics query is done the adding procedure
resets the windowModel and tries to open a dialog (it doesn't succeed
because there is already an open one). But when the user clicks ok on the
add-nic dialog the GuideMe model uses the new window model (with default
values) set after the nic query was done and not the model with which the
dialog was originally opened.

The solution is to reset nics and networks before the starting the queries.

Change-Id: I0445c014831967524f23ae5847b388354b0e7555
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>