engine: Time zones not ordered well (#850696)
https://bugzilla.redhat.com/850696

WebAdmin -> Virtual Machines tab -> new server/desktop ->
Initial Run -> Time Zone
The content of this check box is not ordered well if the
selected VM operating system is windows.

The problem was in GetTimeZonesQuery which had the
TimeZoneTimeExtractComperator not correctly implemented:
The compare method returned only 0 and 1, but never -1,
even when the second arg was smaller than the first one.

This worked in JDK6, but the JDK7's Collections.sort()
checks if the comparator is implemented correctly, and
if not, it does not sort the list at all.

Fixed by fixing the compare method to work according to
the contract.

Change-Id: Iecae3f868d6b3bc6663459367d1bd424b920a3fb
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>