core: Fix quota calc for stateless (#848107)
https://bugzilla.redhat.com/848107

The quota calculation for stateless vm runs was wrong and the error
message shown in the UI was uninformative.

The calculation was fixed - the size of storage consumed from the quota
is the size of the snapshot taken in order to run in stateless mode.

The informative error messages are forwarded from the backend.

Added a release for storage quota when a vm is stopped or removed.

The caching of the quota is actuly dropped - so the next time it will be
called it will be recalculated from the DB.

Change-Id: I4fa99abecfb48a1fc04b7906be8167846ecc629e
Signed-off-by: Ofri Masad <omasad@redhat.com>