engine: Cluster with Virt and Gluster enabled - configurable
A new configuration property is added vdc_options named 'AllowClusterWithVirtGlusterEnabled'.
It is to determine whether the user is allowed to create
a cluster with both Virt and Gluster services enable or not.

It will have the default value as true, so that
user can have a cluster with Virt and Gluster enabled.

Required entries are added to engine-config.properties.

Change-Id: I63c501501545f35ee445861f71ec3c7197d66e06
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>