packaging: Move defaults out of service script
We used to have the defaults values for local engine configuration
files hardcoded in the service start/stop script. This means that
other tools need to duplicate these values. This change moves the
default values to a separate plain file that can be easily
consumed by all kinds of programs.

The file will be stored in /usr/share/ovirt-engine because it is
*not* a configuration file.

Change-Id: I062bb033bf20db9a0d68b8286151345a3653cb03
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>