core: remove check for permissions on pool on detach user from vm (#850741)
https://bugzilla.redhat.com/show_bug.cgi?id=850741

When detaching user from pool-vm, there is a check that the user has
permissions on the pool.

this check is not correct since user may have lost his permissions on
the pool before stopping the vm, this check was removed.
anyway user is detached from vms in this pool so it has no meaning.

Change-Id: I3bfb1a741c6be33427e94d944a8c41d3553cba80
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>