core: setting proper timeout mechanism for GetRootDSE (#844733)
https://bugzilla.redhat.com/844733

This is a first patch in series of patches that deals with
a problem that search users reaches a timeout.
In this patch, a proper timeout mechanism is used.
Although it was assumed that changing LdapQueryTimeout
to a low value (i.e - 3 seconds) will yield a timeout after
the provided amount of time, in fact it took longer for
GetRootDSE to timeout.
This patch fixes this issue.

Change-Id: I59e875997b904cbdc058a45efd640f5bf1fc6579
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>