core, packaging: Don't use JBoss default ports (#831193)
https://bugzilla.redhat.com/831193

This patch changes the default ports used by the engine so that
they don't conflict with the default ports used by JBoss. The port
numbers proposed are the following:

  http: 8700
  https: 8701
  ajp: 8702
  remoting: 8703
  txn-recovery: 8704
  txn-status: 8705

Why these port numbers and not others?

1. They are not allocated according to /etc/services.

2. They are reasonably low.

3. The 8700-8731 range is not allocated by IANA, so we could
request them for ovirt-engine, to reduce the chances of conflicts
in the future.

4. 8700 is reasonabley easy to remember.

5. They are reasonably far away from the ranges used by VNC and spice
displays of virtual machines.

Change-Id: I3c0ffcab66fe9ae6f3dcb932040feb4421e18bf6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>