Revert "engine: Obtain old network name from correct NIC (#849971)"
This reverts commit d9472bf9135f5d230603da30dc9d1ee65fcdf947.

Commit b06f11509e488f2888ed4044081f5041d00fa93e introduced several
regressions. It designed to calculate the network name by the interfaces
being sent (in order to prevent remove of network not related to the
modified interfaces). However it fails to resolve several cases in which
the sent interface doesn't contain the network name (multiple VLANs over
bond where on backend side assumption can't be made).

Change-Id: I853fcd2400ab131f687d798a4cff230a99294c8f
Signed-off-by: Moti Asayag <masayag@redhat.com>