webadmin: Edit pool hangs when the pool has no VMs (#852297)
https://bugzilla.redhat.com/852297

When the pool has no VMs, it makes no sense to have the pool
itself.

This patch adds a behavior, that when the user detaches all the VMs
from the pool, show a message to the user, and when the
user approves this operation, it deletes the pool itself.

The pool is deleted in the RemoveVmFromPoolRunner using the
RemoveVmPoolCommand when there is no more VMs left in the
pool and the operation is explicitly enabled by
removePoolUponDetachAllVMs parameter on the
RemoveVmFromPoolParameters.

Change-Id: I3b8f47e6dd5ce8149ba78c7c25455997b3bf62eb
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>