core: RemoveImageDisk - race when updating snapshots ovf (#828192)
https://bugzilla.redhat.com/show_bug.cgi?id=828192

When removing an image disk it should be removed from all the snapshots
that contain it. The removal from the vm snapshots includes an
update to the snapshot ovf (saved in the DB). The update is a
read-update-write operation, so when two or more disks are removed from
the same snapshot a race condition might occur.
This patch adds a lock on the snapshot when performing the operations on
the snapshot ovf to prevent the race condition.

Change-Id: Iccb44f1aa9d204477955343167133849a4146753
Signed-off-by: Liron Aravot <laravot@redhat.com>