bootstrap: node id: consider bonding MAC addresses as well
The node id is constructed using the lowest MAC address.

If bonding is applied after initial node id was taken, it may shade the
minimal address.

This change adds to the MAC addresses of bonding interfaces to the MAC
address addresses to consider.

Sync with vdsm, consider only physical devices (has /device).

Change-Id: I31ce43972fad77c716aff915d29d6589ba205280
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>