packaging: Find and check JVM during setup (#834436)
https://bugzilla.redhat.com/834436

We used to have a hardcoded setting in the engine.conf.defaults file for
the location of the Java virtual machine. This patch changes the
engine-setup tool so that it will try to find a suitable JVM (suitable
means OpenJDK 1.7 at the moment) automatically. The location of the JVM
will then be written to the JAVA_HOME configuration parameter in the
/etc/sysconfig/ovirt-engine  file.

Change-Id: Ib279ea8111825ec35a25f9ba6ee37b481ed4354f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>