core: checking quota even if can-do fails (#846777)
https://bugzilla.redhat.com/846777

there is a problem in quota flow, that I validate and set (set is the wrong part here)
quota even if the command fails. that causes a lot of problems when the command fails
and the quota get validated, and the use of that command is getting cached...
therefore, the validate & set quota will be invoked iff all the command will go on to
the execute part.

Change-Id: I0f4dd77582985f521ef0009b421cbd0e83466b57
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>