core:Enable update for storage domain name(#851835)
https://bugzilla.redhat.com/show_bug.cgi?id=851835

Update storage domain name fails since lastTimeUsedAsMaster property is
sent from the client is not synchronized with the DB value.
This is because this field is not reflected to the user and it is
internally used in the engine.

The proposed fix is to update the field with the old value from the DB
before update the storage domain in the DB.

Change-Id: I77de69fb8bd28a3c2c1ec1a9be576cb7b4e9dd69
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>