webadmin: Add validation for create/edit pool (#849426)
https://bugzilla.redhat.com/849426

The VM name length is limited. When the new pool
is created, the names of the VMs in the pool is
limited as well. When the limit is, say 6 chars,
then when you create a pool with name like 'pool'
with 10 VMs in it, the names of the VMs will be
pool-1, pool-2,..., pool-10. But the pool-10
is 7 chars long and the creation of such a VM fails.

This patch adds a validation to the new/edit
pool dialog which prevents the user from doing
this. The behavior of the validation is the
following:

New Pool:
The name field gets validated with a message like:
'The max allowed name length is X for Y VMs in pool'

Edit Pool:
The 'Increase number of VMs in pool by' gets validated
with a message like:
'The max allowed num of VMs is X when the length of the pool name is Y'
Where the X is 'num of vms already assigned' + 'num of
VMs by which the pool should be increased'

Change-Id: I66f765ceb0567beb006d1ef4f45db3902ef95ea6
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>