webadmin: [SetupNetworks] detach unmanaged net causes NPE (#848354)
https://bugzilla.redhat.com/848354

Fixing bug 848354- caused NPE when detching unmanaged network.
The fix was- that when detaching unsync network the is_vm, vlan-id,
and mtu parameters should be set to the values defined in the dc
logical network level. The problem is that unmanaged network also
marked as unsync. And on detach it just have to be deleted, no need
to set the dc parameters.

Change-Id: Ic63105d7e352845d795fd99f4099c812723bca6e
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>