webadmin: [setupNetworks] Unsync net params can be changed (#848354)
https://bugzilla.redhat.com/848354

The bug was fixed in a previous commit, but the fix worked just in
case the "is to sync" was checked an unchecked without closing the
dialog.
If the edit network dialog is colsed and then opened again and
the "is to sync" is changed back to false. The original "before sync"
values should be reverted- this didn't work in the prevoius commit.


Change-Id: Ia00641b123f109d443d98e2c99e727099c2a9d16
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>