engine: Prevent move to empty quota (#845070)
https://bugzilla.redhat.com/845070

When a user was moving a disk to a domain with no defined quota (when
quota is enforced) he got an uninformative error.

The ability to move to a domain with no quota was blocked using the
validate mechanism. When quota is enforced the user must choose a domain
with a defined quota.

The massage was fixed to inform the user that the selected domain has no
assigned quota (in case a quota is available but not relevant).

Change-Id: Iab0baf6f583f06197286a976211bada4a0654cb2
Signed-off-by: Ofri Masad <omasad@redhat.com>