core:Initialize host name for CDA message(BZ846771)
https://bugzilla.redhat.com/show_bug.cgi?id=846771

the following patch propose a fix for the scenario when lun disk is
being extended and the host gets network exception while trying to
connect the storage domain.
This cause the CDA message not to be set with the host id that was not
manage to connect to the domain.

The solution is to Wrap the VDS calls with try and catch so failedHost list
can be set with the problematic host before throwing an exception.

Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Change-Id: I3aa40d4671dc99c2126c131fda36c25d8ed4826e