engine,webadmin: Host pinning is ignored on Template creation (#845570)
https://bugzilla.redhat.com/845570

VM main tab -> select VM which has host pinning set up -> make template

The template will not have the host pinning settings set.

The problem was, that this has been ignored on
frontend, backend and DB function level. Added it everywhere.

In user portal the host pinning will not be copyed, because
the frontend will ignore it (UserPortalListModel). This is
correct - the host pinning has to be completely ignored in UP.

Change-Id: Ie341ff1aa9c8e4bc9cbc4ab05b55664a7e0fd416
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>