core: CommandsBase: refactor task creation
In order to create async tasks, CommandBase expects inheriting classes
to override the ConcreteCreateTask method. For classes that do not
overwrite it, an UnsupportedOperationException is thrown.

However, most places that override this method contain the exact same
code, where the only difference is the type of the task created.

This patch suggests a new mechanism so that each inheriting class will
just need to specify what type of task it requires. For commands that do
not specify the type, an UnsupportedOperationException will still be
thrown.

This patch contains the refactoring in CommandBase, the first usecase
for this mechanism as an example (using CreateSnapshotCommand) and a
test case that proves that the behavior has not changed.

If this patch is approved, subsequent patches will apply the new
mechanism to further commands.

Change-Id: I5a4c77067d43ba5c2dbaf46ff855ac14d29de2ff
Signed-off-by: Allon Mureinik <amureini@redhat.com>