core:Hot Plug Nic/Disk creates incorre...(#855651)
core:Hot Plug Nic/Disk creates incorrect duplicate devices

https://bugzilla.redhat.com/show_bug.cgi?id=855651

The Hot Plug Nic/Disk commands does not set the deviceId field.
When we get run time info from VDSM, we are looking for the deviceId
field, it is is not found, the device is assumed to be an unmanaged
device reported to us as a new device by VDSM (i.e. does not exists in
 DB yet)
This will result in a duplicate of each Nic/Disk that is marked as
managed device with another device (with different GUID) in vm_device that
is marked as an unmanaged device while the last (and incorrect device)
will get the address and alias field updates.

This patch sets the deviceId field in Nic/Disk Hot Plug command, since this
field is not touched by VDSM and returned as is, we can match the
returned hot-plug device when updates arrived with current hot plug
devices and update the information instead adding new devices.

Change-Id: Idac67477c800339b0c48f86a3281c90f3fd4ed7b
Signed-off-by: Eli Mesika <emesika@redhat.com>