webadmin: MTU config should be disabled and blocked for 3.0 DC (#854496)
https://bugzilla.redhat.com/854496

A pervious commit fixed the bz, but there was a little problem.
The "MTUOverrideSupported" config value request is async and in some cases
can be retrieved after the dialog and its ok and cancel buttons are loaded.

In this case the user can change the mtu before the field is disabled.

Change-Id: I33d7c24b40ff5acd3ee1249387a93f48f632b3c4
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>