core: remove unmanaged devices when ...(#846696)
core: remove unmanaged devices when run-once VM is stopped

https://bugzilla.redhat.com/846696

When a VM that has an attached CD or Floppy is running with run-once a
duplicate unmanaged copy of the CD/Floppy is created.
When running this VM again, both copies of the Cd/Floppy(the original
managed one and the unmanaged one created by the run-once) will be sent
to VDSM causing an exception in libvirt since only one CD/Floppy is
currently supported.

This patch handles removal of managed devices in StopVmCommandBse in the
same manner as stateless VM is handled.

Change-Id: I1a34e93d0c9ef401afdb8eded95782cbf859dda6
Signed-off-by: Eli Mesika <emesika@redhat.com>