core:Import VM should not be blocked when collapse.(#766657)
https://bugzilla.redhat.com/766657

Import VM to block domain should be blocked only for
images which have sparse volume type and RAW volume format.

Block storage domains do not support RAW sparse volumes, so those disks
can not be imported to ISCSI domain.
If user will choose to copy collapse the VM images, he can choose to
override the disks configuration to be cow sparse or raw preallocated,
which are supported by the target block storage domains, and the VM can
be imported.

The proposed fix, performs the validation on the configuration, in two cases:
1. When user sends copy collapse indication as true, and chooses a valid
configuration for volume format and type for storage domain.
2. when user imports a VM with valid disks configuration without copy collapse.

Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Change-Id: Ic6f4b762fbd9770c5157fee6b24373fa9a0dca10