engine: Adding gluster volume virt options to configuration
While optimizing a gluster volume to be used a virt store,
there will be 3 options needs to be set named 'group', 'storage.owner-uid'
and 'storage.owner-gid'.The values for these options will be stored in the configuration.

Default values will be 'group'='virt', 'storage.owner-uid'='36'
and 'storage.owner-gid'='36'.

Here we are representing vdsm uid as 36 and kvm gid as 36,
as these Ids might change in the future, we are keeping them in the configuration.

Change-Id: I4b0bed1fd2d904c7261f547907a1b3ec34c6afcb
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>