core: audig_log table is not cleaned ...(#856009)
core audig_log table is not cleaned properly, causing rhevm database to grow excessively(#856009)

https://bugzilla.redhat.com/856009

Problem was that the DeleteAuditLogOlderThenDate removed only records
with processed=true.
this flag is used by rhev-notification tool to mark which events were
handled (by sending events to subscribers) , so , the default of this
flag is false and after rhev-notification tool handles it , it is set to
true

Sine the DeleteAuditLogOlderThenDate removes old data (more than 30 days
by default) there is no need to relate to this flag and it should be
removed from the SP condition.

Change-Id: Id1e5d06751d672aa1f40f5a395234f71597289f4
Signed-off-by: Eli Mesika <emesika@redhat.com>