core: error message when deactivating SD with active VMs (#843407)
https://bugzilla.redhat.com/843407

canDoAction() of DeactivateStorageDomainCommand runs a stored procedure
to find VMs whose status should prevent deactivation.  This patch
renames the SP and associated DAO methods to reflect what it is really
searching for (non-"Down" VMs), as well as changes the error message to
accurately reflect the cause of the failure.

Change-Id: I3ce9ce378ffeed980af508e6556f9d844a3e07bd
Signed-off-by: Greg Padgett <gpadgett@redhat.com>