core: add locks to VmTemplate commands (#761515)
https://bugzilla.redhat.com/761515

Currently the different VmTemplate commands make use of DB lock only,
which causes to different race conditions (as the lock is check and only
then acquired - non atomic operation). This patch adds to the different
VmTemplate commands the use of in-memory lock and canDoAction() checks
in order to prevent those race conditions and possible corruption.

NOTE: this patch doesn't add the use of locks to the different export
domain VmTemplate related commands, the reason for this is that the
exported template is not present on the engine DB so a lock on that
export object cannot be done.

Change-Id: I2d5b6518bbf19077ae5822b9359a6147c87d0f46
Signed-off-by: Liron Aravot <laravot@redhat.com>