core: Upgrade status domain
Because of change in behaviour at error flows, all
statuses of storage domains which were InActive should
be moved to Maintaince status.
If status of domain is InActive it means that domain was not disconnected
and decativated by engine.
The reason for these is because before a change InActive and
Maintaince status has similar behaviour at the system.
That's why wea re moving all inactive domains to Maintaince status, in order
to mark them as disconnected and deactivated

Change-Id: Ibf2316023849648ecc8bee45061ccae0ded5e120
Signed-off-by: Michael Kublin <mkublin@redhat.com>