core: Template stuck in Locked status ...(#846324)
core: Template stuck in Locked status after restart of VDSM during template import(#846324)

https://bugzilla.redhat.com/846324

Problem was that stop task command was sent to VDSM even if the status
of the task is unknown or aborting causing VDSM to stop the task without
cleaning and leave garbage behind.

The patch adds a condition to SPMAsyncTask:stopTask that checks if the
task is not in unknown or aborting state before attempting to send a
StopTask command.

Change-Id: I75947f44448ce07f36e295a520e373ad6e7546fe
Signed-off-by: Eli Mesika <emesika@redhat.com>