core: Change auto-recovery code for storage domains
Because of change which was done at error storage flows
these is no need for activating of storage domain in auto-recovery, when domain will be
seen by vdsm it will be automaticlly switched to Active status.
These is possible because of no more disconnect and deactivate is
send when domain in error state.
At any case we will try to reconnect storage domain if it was disconnected.
The status of domain to Active will be switch by vdsm

Change-Id: I04f25dca8888b19711a840aed150227437972d65
Signed-off-by: Michael Kublin <mkublin@redhat.com>