core: Changed illegal error msg to locked(#840304)
Replaced VM in illegal state error message to locked VM in
ImagesHandler.PerformImagesChecks() when the VM is indeed locked.

The specific flow of actions to introduce the wrong message was
Trying to export a VM while committing a preview.
This yielded a statement that the VM's state is illegal while it should
only say that it's locked.
This should correct other flow of events as well.

ACTION_TYPE_FAILED_VM_IS_LOCKED message is added and used in the appropriate
place
- ImagesHandler.PerformImagesChecks() - if the VM is locked and disks
are not.

Change-Id: I2edb7b8c946589f3bbad251d533704fe856bc175
Bug-Url: https://bugzilla.redhat.com/840304
Signed-off-by: Vered Volansky <vvolansk@redhat.com>