webadmin: importing/removing newly detected gluster hosts
An alert will be shown in the Cluster General tab
when there new hosts which were added to the cluster through
gluster cli(gluster peer probe) and they are not available in the ovirt engine.

On clicking of the alert, a dialog will be shown with the new hosts
along with their fingerprints. User will provide the passwords for them
and save.

Change-Id: I7afc2823d59c510d6e6aabd78e9be1d3c2ed082e
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>