core:  Auto-Recovery should check whether getVdsStats returns 'lastCheck<60' before it proclaims host as up (#844438)
https://bugzilla.redhat.com/844438

The described bug is a private case of regular activate host operation.
The problem is that the condition for moving host to NonOperational is different for those that move host to Active state.
Solution is during IniVdsOnUpCommand perform getVdsStats in order to get info about storage domains that host can see,
if all Active domins in cluster is seen by host it will be moved Active, otherwise it will be left as NonOperational
with reason: STORAGE_DOMAIN_UNREACHABLE

Change-Id: Id88ef24829d5fae61dddd34f0265ae3132672783
Signed-off-by: Michael Kublin <mkublin@redhat.com>