webadmin: Apply should be disabled when editing Network params.
https://bugzilla.redhat.com/859074

Apply in Edit Network dialog should be diasbled if editing params is
enabled. So, if the network is attached to any cluster the apply is enabled
until all the clusters are detached.
After all the cluster are detached and editing the network properties is
permitted the apply should become disabled.


Change-Id: I4139c1adbbc23d137e76419a82b377c6b9c2bf0b
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>