engine: upgrade.sh for upgrading multiple DBs
Currently, the upgrade.sh script creates a md5sum file, which is used
for comparing with md5sum of latest files to identify whether an
upgrade is required. However this file name is not db-specific, and
hence upgrading one DB makes it think that all other DBs are also
upgraded.

Modified the script to create the md5sum file with a DB-specific name
so that all DBs can be upgraded by passing the -d option to the script.

Since the name of the md5sum file is changed now, and it can be
different for different databases, updated the .gitignore file to
ignore all such files.

Change-Id: I63f3c34a822508d266a4b3e1582d7e9539d08c39
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>