webadmin: Add quota usage tooltip (#848028)
https://bugzilla.redhat.com/848028

A tooltip added to the usage column in the Quota>storage sub-tab to
explain the differences between that value and the value shown in the
Storage>VMs sub-tab.

The value in Quota>storage sub-tab uses the max growth potential of a
disk to calculate the quota consumption.

The value in Storage>VMs sub-tab shows the actual current use of the
disk.

Change-Id: Ic3f3c0243650e9822509c1249f0f6ca9c512d543
Signed-off-by: Ofri Masad <omasad@redhat.com>