core: task manager cache is not cleared..(#854308)
task manager cache is not cleared after getting error that task does not exist from vdsm on stopTask(#854308)

This patch insures that async tasks found in engine and not found by
VDSM will get the 'unknown' state.
The case in which there may be no SPM temporally as a result of a
failure is handled differently in AsyncTaskManager and
SPMGetAllTasksStatusesVDSCommand.

Previous code sets all tasks that are found by engine (in cache) and not
found as VDSM as running. This prevented cleanup of unknown tasks from
the cache as described in BZ.

https://bugzilla.redhat.com/854308

Change-Id: I979cb7d2273c2eaaa2866dcb4a5f760f9266e11e
Signed-off-by: Eli Mesika <emesika@redhat.com>