core: spm and master domain maintenance actions cause wrong error (#853704)
When failed to acquire lock, the error message is misleading, as it
clains that same action is in progress, but it is not the case.
The real case is that there are two related operations that are occuring
at the same time, so the user should wait for one to finish, before
initiating the second one. This patch fixes this error message.

Change-Id: I183ab61e1078c18b94e0faab5f1ba5074b8c9e8d
Signed-off-by: Oved Ourfali <oourfali@redhat.com>