webadmin: Using editable combo for volume option key
In the Volume option dialog, the keys will be shown in a
combo, the user can either select a key from the list or
type his own key and provide the value for it.

'user.cifs' option is added explicitly to the key list
as it will not available in the original key list.

Change-Id: I48ac5232d95dc01b9f38160c00e37c02dcd44f0a
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>