core: Removing TransactiveAttribute and appropriate logic
The following attribute was used in order to perform a rollback at case
of failure inside Vds commands.
The following approach was used only at one command AddNetworkVDSCommand correctly
and was used incorrect at IrsCreateCommand and ConnectStorageServerVDSCommand.
Also, we should not perform a rollback inside of vds command , a rollback should be
done from BL commands or any other BL manager.
Also, the Vds commands should be run outside scope of transaction, it is mean that
the transactional rollback is not relevant for them.

Change-Id: I12638ee4c27d8548fc5e1a3212ec4541507fd023
Signed-off-by: Michael Kublin <mkublin@redhat.com>