core, webadmin: Use NFS v3 as default
Change NFS version to v3 as default in UI, and add 'auto' param support
to vdsm in ConnectStorageServerVDSCommand.  This allows pre-3.1 storage
pools to keep using v3 after upgrading (in which case protocol_version
is not passed to vdsm), and gives the option of using NFS version auto
negotiation for SDs in newer pools.

Bug-Url: https://bugzilla.redhat.com/871534
Change-Id: I831ecc7050a5487d5365efb94342a3c170dddc6c
Signed-off-by: Greg Padgett <gpadgett@redhat.com>