packaging: Make DB connection pool configurable
Currently the only way to change the database connections pool size is
to modify the /usr/share/ovirt-engine/service/engine-service.xml.in
file. But that file is not marked as configuration so changes are lost
when the package containing it is updated. This patch introduces two new
optional parameters in /etc/sysconfig/ovirt-engine to control the size
of the pool:

  ENGINE_DB_MIN_CONNECTIONS=1
  ENGINE_DB_MAX_CONNECTIONS=100

The default values are 1 and 100 as shown above.

Bug-Url: https://bugzilla.redhat.com/856273
Change-Id: I607e74df406128c7fd9b2e1496b3efa08e11fbac
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>