core: truncate white list upon upgrade
As white-list data is internal configuration
and is not populated due to user activity, it is
possible to truncate the object_column_white_list_table per
upgrade (as post upgrade step).
This fix was introduced due to a scneario in which sshFingerPrint
key was added vds_static (and vds view) and due to the reentrant nature
of add_object_colmn_white_list_table, it was not added to white list,
as the check of "if exists" was done on the whole table.
It is possible to perform "if exist" check per column, but it looks like
an overhead , as explained above

Change-Id: I94f3c474c0209a9049d1d32487d049e1e161214a
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>