engine: vm doesn't boot from cd when CD set in "edit" (#856806)
https://bugzilla.redhat.com/856806

Virtual Machines tab -> select VM -> Edit -> Boot Options ->
select CD
Than setup boot order to First device: CD, second NONE.

The result is, that the VM will not boot from the CD attached.

The problem was in RunVmCommand.AttachCd() which did set up the
correct CD path only, when the
getVm().getboot_sequence() == BootSequence.CD

The problem is, that this condition is only true, when the
boot sequence is first the Hard Disk (C), than the CD (D).
The condition should be, that when it contains CD (D).

Solved by adding the BootSequence.containsSubsequence method
(the tests in BootSequenceTest) which checks if the specific
sequence is a subsequence of the sequence.

Change-Id: I61d39f8559bcffdc154b0b2c01d5303cf2ff1e1a
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>