engine: Run Once dialog ingores emtpy boot options (#857848)
https://bugzilla.redhat.com/857848

When you set the linux boot options in Edit dialog,
than in Run Once dialog remove this options, the run once will
use the boot options set up in the Edit dialog.

The problem was in RunVmCommand.refreshBootParameters method
which expected, that when the parameter is empty, it is run VM,
when it is not empty, it is run once VM, which is not true in
this case.

Fixed by moving the run once specific code to the RunVmOnceCommand.

This patch contains also the following refactorings:
- the refreshBootSequenceParameter inlined to the
  refreshBootParameters
- setLastStartTime removed from the refreshBootSequenceParameter
  and moved to places where the refreshBootSequenceParameter
  has been called

Change-Id: I32e59783d9c4d8e8b61ec96c6bae5577fc59756c
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>