webadmin: 4.9.6 vdsm version is considered as 3.0 instead of 3.1 (#856507)
https://bugzilla.redhat.com/856507

All 4.9 vdsm versions are considered as 3.0 except 4.9.6.
In the webadmin there was a bug, and 4.9.6 was also considered as 3.0.
It caused the guide me to avoid adding 4.9.6 host to 3.1 cluster.

Also, when selecting host in Cluster guide me and clicking cancel,
and then clicking on select host again there was an exception because the
dc d.m was reseted when canceling the selecet window.

Change-Id: I216c78c6882c9ef18d28950b5889e8d9793f3b4e
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>