restapi: Failed to attach vlan over bond
When attempt to attach a vlan over bond, a NPE is thrown on SetupNetwork
command since the interfaces aren't provided with their information.
Instead, the interface list contains a shallow interface with its id and
bond-name only.

The patch verifies that a slave is exists and assure it is sent with its
data as obtained from engine.

Change-Id: I3f05a11b91e2fd5a0367c0be3904b97d5dcb190e
Bug-Url: https://bugzilla.redhat.com/863105
Signed-off-by: Moti Asayag <masayag@redhat.com>