core: search for disks on specific domains
Fix exception due to bad join condition when searching for disks on
specific storage domains.

Before the regression, this worked by auto-completing disk searches
with 'domains' and other objects with 'domain' (note the plurality)
and using different views for each.  Instead of reintroducing this, a
new join was added which connects the views used for disks and domains.

Change-Id: I2a1467a9a5a6ffdcbaf388ce06f43eb6f9f34ab5
Bug-Url: https://bugzilla.redhat.com/861301
Signed-off-by: Greg Padgett <gpadgett@redhat.com>