bootstrap: allow customize bridge name for standard sequence
For gluster integration, we need bootstrap to be able to accept
management bridge name within bootstrap and not relay on vdsm build time
settings.

As gluster only performs standard host bootstrap (as opposed to
registration), the customization will be available only for this
sequence.

In future the registration sequence will be altered to match the standard
sequence.

Change-Id: I373d921fbecbd2fee79c5a430b428f7f20246055
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=863718