webadmin: Adding Gluster Services Sub Tab to Cluster
A new sub tab named 'Services' is added to the Cluster Tab.
This sub tab will be shown only if the cluster supports Gluster.

NFS, SHD are the two different type of services. Hosts in the
Cluster will have NFS, SHD services running. A filter panel is
added to filter the results based on Service Type and host address.

Change-Id: Iea624f12a07f4aa9b1c6ffa1548aaafb16a00a7a
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>