core: Removing reschedule of monitoring of vdsm by another vdsm
The following patch will remove a reschedule of monitoring of vdsm, from
other thread of monitoring.
Such behaviour can cause to spending all quartz thread on reschedule of monitoring of the
same vdsm.
The use case:
1. We have 10 vdsms and 10 vms that are migrating to same vdsm
2. Each vdsm can schedule additional refresh of destination vdsm
3. Result : we has 10 quartz jobs that are trying to monitor the same vdsm and additional job
   which is regular monitoring (runs each 2 seconds)
It is obvious that under load and in more big environment we will have a timeouts when all
quartz threads will be busy by monitoring of the same vdsms

Change-Id: I5534b3b627e992efbce152935ad207c978a7feed
Signed-off-by: Michael Kublin <mkublin@redhat.com>