core: Add support for iLO2 and iLO4 as...(#837907)
Add support for iLO2 and iLO4 as a fencing options(#837907)

ilo2 is mapped to ilo and ilo4 is mapped to ilo3.
The change is only in configuration and is backward compatible.

This patch does not change the naming from ilo<n> to iLo<n> since the
scripts that are called in the fence_agents package use lowercase.
(Actually in bug description this was described only as a Bonus and is
not a part of the required solution)
Also, I didn't apply this bonus since it implies upgrade of current
installations values and I wanted to lower the risk of a regression.

Change-Id: I01af8bdb69b379edb688b1d11d5f020e841ac354
Signed-off-by: Eli Mesika <emesika@redhat.com>