userportal: VMs are not refreshed after edit
In PUP: edit VM, click OK, edit VM again

The values will still be the old ones and you need
to wait until the next refresh cycle to actually
get them refreshed.

The refreshing after edit works like this:
When the popup is closed (the window model is set
to null), the ModelBoundPopupHandler
catches this event and calls the hideAndClearPopup
which calls the force refresh. This ensures,
that after the values are edited, the screen
contains the updated data.

The problem was, that the UserPortalListModel
called the cancel() method right after the
command to update the values has been sent to the server.
Before the call returned, the force refresh
tried to get the updated data, but they
was not necessarily updated yet.

Changed to call the cancel() after the
specific commands returns to ensure that the
data are actually modified.

Change-Id: If37dfffaafd72e385dfed03b3aafbd07cd82ba2d
Bug-Url: https://bugzilla.redhat.com/864059
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>